feat: Add method delete to VehiclePhotoService
This method accepts two arguments: a Long and a UserDetails. It queries the resource by using the id, case it's not found an exception is thrown, case found the method checks if the user is the owner of the resource or has a Role of ADMIN using the methods of userService. If any of these validations fail an exception is thrown by them, if not the resource is deleted using photoRepository's delete method.