refactor: Change getImagesPathByVehicleId return an absolute URL path in its String List
Before this change, the method returned a relative URL. Now it returns an absolute URL path. To achieve this, the method signature was altered to accept a UriComponent builder as an argument. The methods of PhotoController and VehicleController which use this method were also altered to pass this as a parameter.