Refactor: Big Refactoring to avoid N + 1 problem
In vehicle services the following methods were updated: listById, listAll, listAllByUser,save. And added the auxiliar method: getVehiclePhotosId.

In VehicleGetResponseBody was added a constructor which receives a Vehicle object.

In VehicleController updated the methods addVehicle,displayAll and displayById, displayAllByUser.

In VehiclePhotoService updated the method getPhotosUrlByVehicleId.

In PhotoController removes the method getImagesByVehicleId.

This removes all N + 1 queries, and ensures better performance and quicker responses.